Add/remove register operand to/from RegUseList only when needed.
Changed to add/remove a register operand to/from the RegUseList only
when the operand isn't/is in the RegUseList.  This avoids an assertion
triggered by dragging the same MachineBasicBlock out of and into the
MachineFunction for shuffling, because register MachineOperands aren't
removed from the RegUseList when the whole containing MachineBasicBlock
is being removed from a MachineFunction.

Note: this is just a hack.