Optimize prologue/epilogue code for shadow stack.
Optimize prologue/epilogue code for the shadow stack pass, where we fold
the store/load and the add/sub into a single post-indexed store /
pre-indexed load.

This commit also adds a new opcode for pre-indexed load return, which is
mainly used to be recognized as a return.

This commit also refactors the code in RAN.