Fix a bug in Return Address Nullification code.
Fix a bug in the Return Address Nullification code, where we took it for
granted that R12 (IP) would be free to use at every return and hence
used R12 unconditionally for our instrumentation.  It turns out that R12
is not free in cases of a tTAILJMPr using R12 as the target.  This fixes
the code to find a truly free register using the instrumentor utility,
which now works correctly after commits 74133c5, 961d9b1, 005c0a5,
0d3360a, and 496263a.  In cases where no free register can be found, we
spill and restore R4 using tPUSH/tPOP.