Fix a bug in Indirect Call Limiter pass.
Fix a bug in Indirect Call Limiter pass where we only limit the actual
reigster class of indirect function calls but not the formal register
class.  In cases where none of tcGPR (i.e., R0 -- R3 and R12) are
available, a callee-saved register can still be allocated for an
indirect call, even if the virtual register is constrained to tcGPR.

This commit fixes the bug by adding a new opcode for indirect calls,
which constrains the register class to tcGPR so that the register
allocator cannot allocate a callee-saved register for it in the first
place.