Add CLI options for basic block cluster layout randomization.
Add CLI options for basic block cluster layout randomization, which
randomizes the order of basic block clusters in a function.  A basic
block cluster is a list of basic blocks that may fall through from the
first one to the last one.  By randomizing the order of basic block
clusters, we don't need to insert additional jump instructions at the
end of basic blocks that fall through to another.